+++ R/zero_fill.R
@@ -27,7 +27,8 @@
 #' Data <- LTMRpilot(convert_lengths=TRUE, remove_unconverted_lengths=TRUE)
 #'
 #' # For a univariate analysis
-#' Data_filled <- zero_fill(Data, remove_unknown_lengths=TRUE, univariate=TRUE)
+#' species <- c("Clupea pallasii", "Morone saxatilis", "Parophrys vetulus", "Sardinops sagax")
+#' Data_filled <- zero_fill(Data, species=species, remove_unknown_lengths=TRUE, univariate=TRUE)
 #'
 #' # For a multivariate analysis
 #' Data_filled <- zero_fill(Data, remove_unknown_lengths=TRUE, univariate=FALSE)