+++ NEWS.md
@@ -1,3 +1,5 @@
+# LTMRdata 0.6.0
+
 * Removed life stage from the "Taxa" column.
 * Added option to `zero_fill` to filter to specified species, along with more tests to ensure there are zeros to fill.